requirements.txt
# ================================================================
# SignAI - Requirements File
# AI Sign Language Translation System
# Python 3.8+ Required
# ================================================================

# ================= WEB FRAMEWORK =================
Flask==3.0.0
Werkzeug==3.0.1

# ================= DATABASE =================
# Using PyMySQL instead of mysql-connector-python to avoid protobuf conflicts 
PyMySQL==1.1.0

# ================= COMPUTER VISION & AI =================
opencv-python==4.8.1.78
mediapipe==0.10.8
tensorflow==2.15.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
Pillow==10.1.0

# ================= TRANSLATION & LANGUAGE =================
googletrans==3.1.0a0
gtts==2.5.0
pyttsx3==2.90

# ================= UTILITIES =================
python-dotenv==1.0.0
requests==2.31.0

# ================================================================
# INSTALLATION INSTRUCTIONS
# ================================================================
# 
# 1. Create virtual environment:
#    python -m venv venv
#
# 2. Activate virtual environment:
#    Windows: venv\Scripts\activate
#    Mac/Linux: source venv/bin/activate
#
# 3. Install requirements:
#    pip install -r requirements.txt
#
# 4. If you encounter issues with specific packages:
#    - TensorFlow: pip install tensorflow==2.13.0 --no-cache-dir
#    - OpenCV: pip install opencv-python-headless==4.8.0.76 (for servers)
#    - MediaPipe: pip install mediapipe==0.10.3 --no-deps (if conflicts)
#
# 5. For GPU support (optional):
#    pip install tensorflow-gpu==2.13.0
#
# ================================================================
# ALTERNATIVE LIGHTWEIGHT VERSION (if memory/size is an issue)
# ================================================================
# 
# Replace these packages with lighter alternatives:
# - tensorflow → tensorflow-cpu (CPU only, smaller)
# - opencv-python → opencv-python-headless (no GUI)
# - Remove pyttsx3 if offline TTS not needed
#
# ================================================================
# PLATFORM-SPECIFIC NOTES
# ================================================================
#
# WINDOWS:
# - May need Microsoft Visual C++ 14.0 or greater
# - Download from: https://visualstudio.microsoft.com/visual-cpp-build-tools/
#
# MAC:
# - May need to install portaudio for pyttsx3:
#   brew install portaudio
#
# LINUX (Ubuntu/Debian):
# - sudo apt-get update
# - sudo apt-get install python3-dev
# - sudo apt-get install libportaudio2
# - sudo apt-get install espeak (for pyttsx3)
#
# ================================================================
# TROUBLESHOOTING
# ================================================================
#
# Issue: "No module named '_bz2'"
# Solution: Install bzip2 development files
#   Windows: Reinstall Python with all optional features
#   Linux: sudo apt-get install libbz2-dev
#
# Issue: TensorFlow installation fails
# Solution: Try older version or use CPU-only version
#   pip install tensorflow-cpu==2.13.0
#
# Issue: MediaPipe conflicts
# Solution: Install without dependencies first
#   pip install mediapipe==0.10.3 --no-deps
#   Then install missing dependencies manually
#
# Issue: googletrans not working
# Solution: Use the RC version specifically
#   pip install googletrans==4.0.0rc1
#
# ================================================================